Add authorization and authentication in api using JWT

+++ controllers/api/v1/posts_api.js
@@ -20,16 +20,17 @@ module.exports.index = async function(req, res){
 module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
-        // if(post.user == req.user.id){ //.id used instead of ._id as .id will convert id to string
-        post.remove();
-        await Comment.deleteMany({post: req.params.id});
-        return res.json(200, {
-            message: "Post and associated comments deleted successfully"
+        if(post.user == req.user.id){ //.id used instead of ._id as .id will convert id to string
+            post.remove();
+            await Comment.deleteMany({post: req.params.id});
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully"
         });
-        // }else{
-        //     req.flash('error', 'You cannot delete this post');
-        //     return res.redirect('back');
-        // }
+        }else{
+            return res.json(401, {
+                message: 'You cannot delete this Post'
+            });
+        }
     }catch(err){
         console.log(err);
         return res.json(500, {
